pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opencode-telegram-bot"
version = "0.1.0"
description = "Telegram bot client for OpenCode CLI — run and monitor AI coding tasks from your phone"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "2241812"},
]
keywords = ["opencode", "telegram", "bot", "ai", "coding-agent", "openclaw-alternative"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "python-telegram-bot>=21.0",
    "aiohttp>=3.11",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "sseclient-py>=1.8",
    "apscheduler>=3.10",
    "flask>=3.1",
    "openai>=1.60",
    "httpx>=0.28",
    "prompt-toolkit>=3.0",
    "rich>=13.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[project.scripts]
opencode-tg = "opencode_telegram_bot.main:main"
opencode-telegram = "opencode_telegram_bot.main:main"

[project.urls]
Homepage = "https://github.com/2241812/opencode-telegram-bot"
Repository = "https://github.com/2241812/opencode-telegram-bot"
Issues = "https://github.com/2241812/opencode-telegram-bot/issues"
Documentation = "https://github.com/2241812/opencode-telegram-bot/tree/main/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
